requirements.txt
ansicon==1.89.0
bcrypt==4.2.1
blessed==1.20.0
botocore==1.36.6
chalice==1.31.3
click==8.1.8
colorama==0.4.6
editor==1.6.6
inquirer==3.4.0
jinxed==1.3.0
jmespath==1.0.1
psycopg2-binary==2.9.10
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
readchar==4.2.1
runs==1.2.2
six==1.17.0
urllib3==2.3.0
wcwidth==0.2.13
xmod==1.8.1
